pyproject.toml
[project]
name = "newspaper-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "pillow-heif",
]
